backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
tensorflow
numpy
python-multipart
websockets
Pillow
fastapi-cors
pyyaml
scikit-learn
